Fix the error of "boot failed" after the new kernel loaded
Before: In the "readsector" function, "res = (long long)(FSC - > PartitionStart) * SECTORSIZE +
SECTORSIZE * sector", the type is not consistent, when "sector" is large, the result is not correct,
which lead to some errors.
Now: Forced convert all types to the same type, the "readsector" function is OK.
Test: Put a newly compiled kernel into a FAT formated of usb disk, boot into the pmon interface,
user can load the kernel successlly.

Targets: all

+++ pmon/fs/fatfs.c
@@ -992,7 +992,7 @@ int readsector(struct fat_sc *fsc, int sector, int count, u_int8_t *buffer)
 {
 	long long res;
 
-	res = (long long)(fsc->PartitionStart) * SECTORSIZE + SECTORSIZE * sector;
+	res = (unsigned long long)(fsc->PartitionStart) * (unsigned long long)SECTORSIZE + (unsigned long long)SECTORSIZE * (unsigned long long)sector;
 	res = devio_lseek(fsc->fd,	res, 0);
 	res =devio_read(fsc->fd, buffer, (SECTORSIZE * count));
 	if (res == (SECTORSIZE * count)) {
